requirements.txt
python-dotenv
streamlit
phidata
phi
google-generativeai
duckduckgo-search